fixed some castings and some array declatrations that dont work with MSVC

+++ pufferlib/ocean/rubiks/binding.c
@@ -4,11 +4,11 @@
 #include "../env_binding.h"
 
 static int my_init(Env* env, PyObject* args, PyObject* kwargs) {
-    env->N = unpack(kwargs, "N");
-    env->shuffles = unpack(kwargs, "shuffles");
-    env->size = unpack(kwargs, "size");
-    env->max_episode_steps = unpack(kwargs, "max_episode_steps");
-    env->anim_time = unpack(kwargs, "anim_time");
+    env->N = (int) unpack(kwargs, "N");
+    env->shuffles = (int) unpack(kwargs, "shuffles");
+    env->size = (int) unpack(kwargs, "size");
+    env->max_episode_steps = (int) unpack(kwargs, "max_episode_steps");
+    env->anim_time = (float) unpack(kwargs, "anim_time");
     init(env);
     return 0;
 }
